refactor(http): Update return styles
